Major change to how commands are invoked.
Previously, you would do this:

    $ php cli.php aura.framework.run-tests

Among other things, this required that you have a mapping for the command in the config somewhere, and it meant you could not use tab-completion.

This change makes it so that commands are represented in the package structure.  You still write a command class, but instead of invoking it via the cli.php script with the help of the config file, you now write a very short script to invoke the command, and call it directly.

    $ php package/Aura.Framework/commands/run-tests

This tradeoff allow allows you to use tab-completion, and removes the need to map in a config file.

The invoking file should go in the package/commands directory, and looks like this:

    <?php
    require dirname(dirname(dirname(__DIR__))) . '/package/Aura.Framework/src/Command.php';
    \Aura\Framework\Command::exec('Aura\\Framework\\Cli\\RunTests\\Command');

At some point we could add a command that creates this command file for us.